requirements.txt
Django==2.0.4
djangorestframework==3.8.1
pytest-django==3.1.2

